refactor: move modules in DI folder
