+++ packages/lib/di/containers/available-slots.ts
@@ -1,11 +1,12 @@
 import { createContainer } from "@evyweb/ioctopus";
 
 import { DI_TOKENS } from "@calcom/lib/di/tokens";
-import { oooRepositoryModule } from "@calcom/lib/server/modules/ooo";
-import { scheduleRepositoryModule } from "@calcom/lib/server/modules/schedule";
 import { prismaModule } from "@calcom/prisma/prisma.module";
 import type { AvailableSlotsService } from "@calcom/trpc/server/routers/viewer/slots/util";
-import { availableSlotsModule } from "@calcom/trpc/server/routers/viewer/slots/util";
+
+import { availableSlotsModule } from "../modules/available-slots";
+import { oooRepositoryModule } from "../modules/ooo";
+import { scheduleRepositoryModule } from "../modules/schedule";
 
 const container = createContainer();
 container.load(DI_TOKENS.PRISMA_MODULE, prismaModule);
