+++ packages/lib/di/modules/ooo.ts
@@ -1,8 +1,7 @@
 import { createModule } from "@evyweb/ioctopus";
 
 import { DI_TOKENS } from "@calcom/lib/di/tokens";
-
-import { PrismaOOORepository } from "../repository/ooo";
+import { PrismaOOORepository } from "@calcom/lib/server/repository/ooo";
 
 export const oooRepositoryModule = createModule();
 oooRepositoryModule.bind(DI_TOKENS.OOO_REPOSITORY).toClass(PrismaOOORepository, [DI_TOKENS.PRISMA_CLIENT]);
